+++ config/splat.us.stno1.yaml
@@ -60,6 +60,7 @@ segments:
       - [0x34A00, .rodata, e_stage_name]
       - [0x34A34, .rodata, e_skeleton]
       - [0x34A50, .rodata, unk_4DDD8]
+      - [0x34AFC, .rodata, unk_4F57C]
       - [0x34B6C, .rodata, unk_51DAC]
       - [0x34CBC, .rodata, prim_helpers]
       - [0x34CC4, c, first_c_file]
@@ -80,6 +81,7 @@ segments:
       - [0x4D2A8, c, popup]
       - [0x4D5D8, c, e_skeleton]
       - [0x4DDD8, c, unk_4DDD8]
+      - [0x4F57C, c, unk_4F57C]
       - [0x51DAC, c, unk_51DAC]
       - [0x55AF8, c, prim_helpers]
       - [0x56328, sbss]
